add per app details
