+++ AC Widget/Components/InfoTile.swift
@@ -9,17 +9,7 @@ import BetterToStrings
 struct InfoTile: View {
     private var description: LocalizedStringKey
     private var data: ACData
-    private var rawData: [(Float, Date)]
     private var type: InfoType
-    private var color: Color
-
-    @State private var currentIndex: Int?
-    private var graphData: [CGFloat] {
-        let copy = rawData.map { $0.0 }
-        let max: Float = copy.max() ?? 1
-        return copy.map { CGFloat($0 / max) }
-    }
-
     private var currencySymbol: String {
         switch type {
         case .proceeds:
@@ -28,153 +18,92 @@ struct InfoTile: View {
             return ""
         }
     }
-
-    init(description: LocalizedStringKey, data: ACData, type: InfoType, color: Color = .accentColor) {
-        self.description = description
-        self.data = data
-        self.rawData = data.getRawData(type, lastNDays: 30).reversed()
-        self.type = type
-        self.color = color
+    @State private var isFlipped: Bool = false
+    @State private var interval: Int = 0
+    private var lastNDays: Int {
+        switch interval {
+        case 1:
+            return 7
+        case 2:
+            return 30
+        case 3:
+            return data.latestReportingDate().dateToMonthNumber()
+        default:
+            return 1
+        }
     }
 
+    init(description: LocalizedStringKey, data: ACData, type: InfoType) {
+        self.description = description
+        self.data = data
+        self.type = type
+    }
     var body: some View {
         Card {
-            // Temporarily disabled update numbers (GitHub Issue #3)
-            if type == .updates {
-                HStack {
-                    Text(description)
-                    Spacer()
-                    Image(systemName: InfoType.updates.systemImage)
-                }
-                .font(.system(size: 20))
-                .padding(.bottom, 20)
+            if isFlipped {
+                backside
+                    .rotation3DEffect(Angle(degrees: 180), axis: (x: CGFloat(0), y: CGFloat(10), z: CGFloat(0)))
             } else {
-                topSection
-            }
-            graphSection
-                .frame(minHeight: 100)
-            if type != .updates {
-                bottomSection
+                InfoTileFront(description: description, data: data, type: type)
+                    .contentShape(Rectangle())
+                    .onTapGesture {
+                        withAnimation {
+                            self.isFlipped.toggle()
+                        }
+                    }
             }
         }
         .frame(height: 250)
+        .rotation3DEffect(self.isFlipped ? Angle(degrees: 180): Angle(degrees: 0), axis: (x: CGFloat(0), y: CGFloat(10), z: CGFloat(0)))
+        .frame(height: 250)
     }
 
-    // MARK: Top
-    var topSection: some View {
-        HStack(alignment: .top) {
-            if let index = currentIndex {
-                Text(getGraphDataPoint(index).1.toString(format: "dd. MMM.", smartConversion: true))
-                    .font(.system(size: 20))
-                Spacer()
-                if currencySymbol.isEmpty {
-                    UnitText(getGraphDataPoint(index).0.toString(abbreviation: .intelligent, maxFractionDigits: 2), metricSymbol: type.systemImage)
-                } else {
-                    UnitText(getGraphDataPoint(index).0.toString(abbreviation: .intelligent, maxFractionDigits: 2), metric: currencySymbol)
-                }
-            } else {
-                Text(description)
-                    .font(.system(size: 20))
-                Spacer()
-                if currencySymbol.isEmpty {
-                    UnitText(data.getAsString(type, lastNDays: 1), metricSymbol: type.systemImage)
-                } else {
-                    UnitText(data.getAsString(type, lastNDays: 1), metric: currencySymbol)
-                }
-            }
-        }
-    }
-
-    private func getGraphDataPoint(_ index: Int) -> (Float, Date) {
-        if index >= rawData.count {
-            return rawData.last ?? (0, Date(timeIntervalSince1970: 0))
-        }
-        if index < 0 {
-            return rawData.first ?? (0, Date(timeIntervalSince1970: 0))
-        }
-        return rawData[index]
-    }
-
-    // MARK: Graph
-    var graphSection: some View {
-        Group {
-            if !graphData.isEmpty {
-                GeometryReader { reading in
-                    HStack(alignment: .bottom, spacing: 0) {
-                        ForEach(graphData.indices) { i in
-                            Capsule()
-                                .frame(width: (reading.size.width/CGFloat(graphData.count))*0.7, height: reading.size.height * getGraphHeight(i))
-                                .foregroundColor(getGraphColor(i))
-                                .opacity(currentIndex == i ? 0.7 : 1)
-
-                            if i != graphData.count-1 {
-                                Spacer()
-                                    .frame(minWidth: 0)
-                            }
-                        }
-                    }
-                    .contentShape(Rectangle())
-                    .highPriorityGesture(DragGesture(minimumDistance: 20)
-                                .onChanged({ value in
-                                    let newIndex = Int((value.location.x/reading.size.width) * CGFloat(graphData.count))
-                                    if newIndex != currentIndex && newIndex < rawData.count && newIndex >= 0 {
-                                        currentIndex = newIndex
-                                        UISelectionFeedbackGenerator()
-                                            .selectionChanged()
-                                    }
-                                })
-                                .onEnded({ _ in
-                                    withAnimation(Animation.easeOut(duration: 0.2)) {
-                                            currentIndex = nil
-                                        }
-                                })
-                    )
-                }
-            } else {
-                Text("NO_DATA")
-                    .foregroundColor(.gray)
-                    .italic()
-            }
-        }
-    }
-
-    private func getGraphHeight(_ i: Int) -> CGFloat {
-        if i < graphData.count && graphData[i] > 0 {
-            return graphData[i]
-        }
-        if i < graphData.count && graphData[i] < 0 {
-            return abs(graphData[i])
-        }
-        return 0.01
-    }
-
-    private func getGraphColor(_ i: Int) -> Color {
-        var result: Color = .gray
-        if i < graphData.count && graphData[i] > 0 {
-            result = color
-        } else if i < graphData.count && graphData[i] < 0 {
-            result = .red
-        }
-        return result
-    }
-
-    // MARK: Bottom
-    var bottomSection: some View {
+    var backside: some View {
         VStack {
-            HStack(alignment: .bottom) {
-                DescribedValueView(description: "LAST_SEVEN_DAYS", value: data.getAsString(type, lastNDays: 7, size: .compact).appending(currencySymbol))
+            Picker("SELECT_INTERVAL", selection: $interval) {
+                Text("1D").tag(0)
+                Text("7D").tag(1)
+                Text("30D").tag(2)
+                Text(Date.now.toString(format: "MMM")).tag(3)
+            }
+            .pickerStyle(.segmented)
+            ScrollView(showsIndicators: false) {
+                LazyVGrid(columns: [GridItem(.adaptive(minimum: 150))], spacing: 8) {
+                    ForEach(data.apps) { app in
+                        appDetail(for: app)
+                    }
+                }
+            }
+            .contentShape(Rectangle())
+            .onTapGesture {
+                withAnimation {
+                    self.isFlipped.toggle()
+                }
+            }
+        }
+    }
+
+    @ViewBuilder
+    private func appDetail(for app: ACApp) -> some View {
+        Card(alignment: .leading, spacing: 5, innerPadding: 10, color: .secondaryCardColor) {
+            HStack(spacing: 4) {
+                AsyncImage(url: URL(string: app.artworkUrl60)) { image in
+                    image.resizable()
+                } placeholder: {
+                    Rectangle().foregroundColor(.secondary)
+                }
+                .frame(width: 15, height: 15)
+                .cornerRadius(3)
+
+                Text(app.name)
+                    .lineLimit(1)
                 Spacer()
-                    .frame(width: 40)
-                DescribedValueView(description: "LAST_THIRTY_DAYS", value: data.getAsString(type, lastNDays: 30, size: .compact).appending(currencySymbol))
             }
 
-            HStack(alignment: .bottom) {
-                DescribedValueView(description: "CHANGE_PERCENT", value: data.getChange(type).appending("%"))
-                Spacer()
-                    .frame(width: 40)
-                DescribedValueView(descriptionString: data.latestReportingDate().toString(format: "MMMM").appending(":"),
-                                   value: data.getAsString(type, lastNDays: data.latestReportingDate().dateToMonthNumber(),
-                                                           size: .compact).appending(currencySymbol))
+            if currencySymbol.isEmpty {
+                UnitText(data.getAsString(type, lastNDays: lastNDays), metricSymbol: type.systemImage)
+            } else {
+                UnitText(data.getAsString(type, lastNDays: lastNDays), metric: currencySymbol)
             }
         }
     }
@@ -183,9 +112,9 @@ struct InfoTile: View {
 struct InfoTile_Previews: PreviewProvider {
     static var previews: some View {
         LazyVGrid(columns: [GridItem(.adaptive(minimum: 320))], spacing: 8) {
-            InfoTile(description: "DOWNLOADS", data: ACData.example, type: .downloads)
             InfoTile(description: "PROCEEDS", data: ACData.example, type: .proceeds)
-            InfoTile(description: "UPDATES", data: ACData.example, type: .updates)
+            InfoTile(description: "PROCEEDS", data: ACData.example, type: .proceeds)
+                //.preferredColorScheme(.dark)
         }.padding()
     }
 }
